==> notebooks/0-data-cleaning.ipynb <==
import pandas as pd
import numpy as np

### Data Cleaning
* Combine necessary datasets
* Clean `Song` column
* Format `Date` column to reference the decade only
* Finalize by combining all data sets

top_songs_50s = pd.read_csv('../data/raw/top_1950.csv', index_col=0)
top_songs_58_69 = pd.read_csv('../data/raw/top_1958_69.csv', index_col=0)

top_songs_50s.head()

Combining the two data sets

all_50s_songs = top_songs_50s.append(top_songs_58_69, ignore_index=True)
all_50s_songs.shape

# make column names lowercase for ease

cols = all_50s_songs.columns
cols = [c.lower() for c in cols]

# reassign columns back

all_50s_songs.columns = cols
all_50s_songs.columns

Cleaning the `song` column

all_50s_songs

all_50s_songs['song'][0]

all_50s_songs['song'] = all_50s_songs['song'].apply(lambda x: x.rstrip('"'))
all_50s_songs['song'] = all_50s_songs['song'].apply(lambda x: x.lstrip('"'))
all_50s_songs['song'][:50]

to_clean = all_50s_songs[all_50s_songs['song'].str.contains('"')]

to_clean_songs = to_clean['song'].to_list()
to_clean_songs

cleaned = []
for song in to_clean_songs:
    idx = song.index('"')
    new = song[:idx]
    cleaned.append(new)

cleaned

for i,row in all_50s_songs[all_50s_songs['song'].str.contains('"')].iterrows():
    old = row['song']
    idx = old.index('"')
    new = old[:idx]
    all_50s_songs['song'][i] = new
    all_50s_songs['song'][i] = all_50s_songs['song'][i]
    
all_50s_songs[all_50s_songs['song'].str.contains('"')]

Format the `date` column

# make date column just the year
all_50s_songs['date'] = all_50s_songs['date'].apply(lambda x: x[-4:])

# make date column be int dtype
all_50s_songs['date'] = all_50s_songs['date'].astype('int64')
all_50s_songs.info()

# create a function to assign a decade to each song
def assign_decade(year):
    if (year >= 1950) and (year < 1960):
        return 1950
    elif (year >= 1960) and (year < 1970):
        return 1960
    elif (year >= 1970) and (year < 1980):
        return 1970
    elif (year >= 1980) and (year < 1990):
        return 1980
    elif (year >= 1990) and (year < 2000):
        return 1990
    elif (year >= 2000) and (year < 2010):
        return 2000
    elif year >= 2010:
        return 2010
    
all_50s_songs['decade'] = all_50s_songs['date'].apply(lambda x: assign_decade(x))
all_50s_songs

I'll drop the `date` column at the end.

#### Cleaning the other datasets

top_1970 = pd.read_csv('../data/raw/top_1970.csv', index_col=0)
top_1980 = pd.read_csv('../data/raw/top_1980.csv', index_col=0)
top_1990 = pd.read_csv('../data/raw/top_1990.csv', index_col=0)
top_2000 = pd.read_csv('../data/raw/top_2000.csv', index_col=0)
top_2010 = pd.read_csv('../data/raw/top_2010.csv', index_col=0)

Combine the datasets.

all_songs = pd.concat([top_1970,top_1980,top_1990,top_2000,top_2010], ignore_index=True)
all_songs

# all cleaning steps

# make column headers lowercase
cols = all_songs.columns
cols = [c.lower() for c in cols]
all_songs.columns = cols

all_songs

Some songs are missing for some reason, I'll need to find out what the song titles are.

all_songs[all_songs['song'].isnull()==True]

# checking to make sure records near the missing values are correct
all_songs.iloc[160:200,:]

# looks like I messed up cleaning songs that are themes
all_songs.loc[36,'song'] = "Theme from Shaft"
all_songs.loc[160,'song'] = "Theme from Mahogany (Do You Know Where You're Going To)"
all_songs.loc[163,'song'] = "Theme From S.W.A.T."
all_songs.loc[190,'song'] = "Love Theme From A Star Is Born (Evergreen)"
all_songs.loc[202,'song'] = "Gonna Fly Now (Theme From Rocky)"
all_songs.loc[356,'song'] = "Miami Vice Theme"

all_songs[all_songs['song'].isnull()==True]

# for i,row in all_songs[all_songs['song'].str.contains('"')].iterrows():
#     old = row['song']
#     idx = old.index('"')
#     new = old[:idx]
#     all_songs['song'][i] = new
#     all_songs['song'][i] = all_songs['song'][i]

# check
all_songs[all_songs['song'].str.contains('"')]['song']

# checking to make sure I didn't accidentally create missing values
all_songs.info()

# make date column just the year
all_songs['date'] = all_songs['date'].apply(lambda x: x[-4:])

# make date column be int dtype
all_songs['date'] = all_songs['date'].astype('int64')
all_songs.info()

# assigning the decade label
all_songs['decade'] = all_songs['date'].apply(lambda x: assign_decade(x))
all_songs

all_decade_songs = all_50s_songs.append(all_songs, ignore_index=True)

all_decade_songs.info()

all_decade_songs[all_decade_songs['song'].str.contains('"')]

all_decade_songs['song'] = all_decade_songs['song'].apply(lambda x: x.rstrip('"'))
# all_decade_songs['song'] = all_decade_songs['song'].apply(lambda x: x.lstrip('"'))
all_decade_songs['song'][974:]

for i,row in all_decade_songs[all_decade_songs['song'].str.contains('"')].iterrows():
    old = row['song']
    idx = old.index('"')
    new = old[:idx]
    all_decade_songs['song'][i] = new
    all_decade_songs['song'][i] = all_decade_songs['song'][i]

# check
all_decade_songs[all_decade_songs['song'].str.contains('"')]

all_decade_songs.info()

# all_decade_songs.drop('date', axis=1, inplace=True)
all_decade_songs

year_key = all_decade_songs['date']
year_key.to_csv("../data/interim/year_key.csv")

all_decade_songs.to_csv('../data/all_decade_songs.csv', index=False)
