==> notebooks/1-data-feature-collection.ipynb <==
### Data/Feature Collection
Notebook for collecting features from the Spotify API and Librosa.

import pandas as pd
import numpy as np
import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import librosa
import json
import requests
from secret import *
from IPython.display import clear_output

client_credentials_manager = SpotifyClientCredentials(client_id=spotify_credentials['client_id'],
                                                      client_secret=spotify_credentials['client_secret'])

sp = spotipy.Spotify(client_credentials_manager=client_credentials_manager)

all_decade_songs = pd.read_csv('../data/all_decade_songs.csv')
all_decade_songs

Get song objects from Spotify.

def get_missing_url(artist,song):
    '''falls back on the iTunes API to get a 30 sec. preview of a song if Spotify
        doesn't provide one, also assigns a different genre since iTunes uses
        more traditional genres, returns track metadata'''

    artist = artist.replace(" ","+")
    song = song.replace(" ","+")

    try:
        r = requests.get(f"https://itunes.apple.com/search?term={artist}+{song}&limit=1")
        content = json.loads(r.text)
        preview = content['results'][0]["previewUrl"]
        genre = content['results'][0]["primaryGenreName"]
        return str(preview), genre.lower()

    except:
        pass
    
def search_and_extract(track_query):
    '''A function that takes in a song query and returns
    the track id and preview url for that track in a dict.'''

    track_query = str(track_query).strip().replace(" ","+")

    #uses the API to search for a track
    try:
        search = sp.search(track_query, type='track', limit=1, market='US')

        track_id = search['tracks']['items'][0]['id']
        preview_url = search['tracks']['items'][0]['preview_url']
        track_name = search['tracks']['items'][0]['name']
        artist = search['tracks']['items'][0]['artists'][0]['name']
        artist_id = search['tracks']['items'][0]['artists'][0]['id']

        search = sp.artist(artist_id)
        genre_list = search['genres']
        
        track_data = [track_id, preview_url, track_name, artist, artist_id, genre_list]

    except:
        track_query = str(track_query).split(",")
        name = track_query[0]
        search = sp.search(name, type='track', limit=1, market='US')

        track_id = search['tracks']['items'][0]['id']
        preview_url = search['tracks']['items'][0]['preview_url']
        track_name = search['tracks']['items'][0]['name']
        artist = search['tracks']['items'][0]['artists'][0]['name']
        artist_id = search['tracks']['items'][0]['artists'][0]['id']

        search = sp.artist(artist_id)
        genre_list = search['genres']


        track_data = [track_id, preview_url, track_name, artist, artist_id, genre_list]


    return track_data

all_decade_songs_df = all_decade_songs.copy()

all_decade_songs_df['track_id'] = np.nan
all_decade_songs_df['preview_url'] = np.nan
all_decade_songs_df['track_name'] = np.nan
all_decade_songs_df['artist_name'] = np.nan
all_decade_songs_df['artist_id'] = np.nan
all_decade_songs_df['genres'] = np.nan

error = []

for i,row in all_decade_songs_df.iterrows():
    artist = row['artist']
    song = row['song']
    
    try:
        query = f"{song} {artist}"
        track_data = search_and_extract(query)
        all_decade_songs_df.loc[i,'track_id'] = track_data[0]
        all_decade_songs_df.loc[i,'preview_url'] = track_data[1]
        all_decade_songs_df.loc[i, 'track_name'] = track_data[2]
        all_decade_songs_df.loc[i, 'artist_name'] = track_data[3]
        all_decade_songs_df.loc[i,'artist_id'] = track_data[4]
        all_decade_songs_df.loc[i,'genres'] = track_data[5]
        
    except:
        
        try:
            query = f"{song}"
            track_data = search_and_extract(query)
            all_decade_songs_df.loc[i,'track_id'] = track_data[0]
            all_decade_songs_df.loc[i,'preview_url'] = track_data[1]
            all_decade_songs_df.loc[i, 'track_name'] = track_data[2]
            all_decade_songs_df.loc[i, 'artist_name'] = track_data[3]
            all_decade_songs_df.loc[i,'artist_id'] = track_data[4]
            all_decade_songs_df.loc[i,'genres'] = track_data[5]
            
        except:
            error.append(f"{song} {artist}")
            
    clear_output(wait=True)
    print(f"{i}/{len(all_decade_songs_df)}")

error

missing_idx = all_decade_songs_df[all_decade_songs_df['track_name'].isnull()==True].index
all_decade_songs_df.loc[missing_idx,'track_id'] = missing_tracks
all_decade_songs_df[all_decade_songs_df['track_name'].isnull()==True]

for i,row in all_decade_songs_df[all_decade_songs_df['track_name'].isnull()==True].iterrows():
    
    search = sp.track(row['track_id'])
    
    try:
        all_decade_songs_df.loc[i,'track_id'] = search['id']
        all_decade_songs_df.loc[i,'preview_url'] = search['preview_url']
        all_decade_songs_df.loc[i,'track_name'] = search['name']
        all_decade_songs_df.loc[i,'artist_name'] = search['artists'][0]['name']
        all_decade_songs_df.loc[i,'artist_id'] = search['artists'][0]['id']

        search = sp.artist(artist_id)
        all_decade_songs_df.loc[i,'genres'] = search['genres']
        
    except:
        pass

    

all_decade_songs_df.iloc[missing_idx,:]

all_decade_songs_df[all_decade_songs_df['artist'] != all_decade_songs_df['artist_name']]

__Notes:__
There is still some cleaning to do. Very few rows have their associated genre, the `artist` and `artist_name` columns don't match, which means they are difference versions. Also, several previews are missing and will need to be tracked down.

all_decade_songs_df.to_csv('../data/all_decades_songs_V1.csv')
