==> notebooks/5-librosa-pipeline-dask.ipynb <==
!pip install dask --upgrade

!pip install distributed --upgrade

import pandas as pd
import numpy as np
import librosa
import time

from dask.distributed import Client
client = Client(n_workers=8)

from dask import delayed, compute

songs = pd.read_csv('all_decades_songs_V3.csv', index_col=0)

def librosa_pipeline(track_id):
   path = f'drive/My Drive/previews/track_{track_id}.wav'
   
   d = {}
   d['track_id'] = track_id

   #load mp3
   y, sr = librosa.load(path, mono=True, duration=30)
   
   #feature extraction
   spec_cent = librosa.feature.spectral_centroid(y=y, sr=sr)
   d['spectral_centroid'] = np.mean(spec_cent)
   
   spec_bw = librosa.feature.spectral_bandwidth(y=y, sr=sr)
   d['spectral_bandwidth'] = np.mean(spec_bw)
   
   rolloff = librosa.feature.spectral_rolloff(y=y, sr=sr)
   d['rolloff'] = np.mean(rolloff)
   
   zcr = librosa.feature.zero_crossing_rate(y)
   d['zero_crossing_rate'] = np.mean(zcr)

   mfcc = librosa.feature.mfcc(y=y, sr=sr, n_mfcc=20)
   for i,e in zip(range(1, 21),mfcc):
     d[f'mfcc{i}'] = np.mean(e)

   chroma = ['C', 'C#', 'D','D#', 'E', 'F', 'F#', 'G', 'G#', 'A', 'A#', 'B']
   chroma_stft = librosa.feature.chroma_stft(y=y, sr=sr)
   for c,p in zip(chroma,chroma_stft):
     d[c] = np.mean(p)

   return d

librosa_features = []

start = time.time()
for i,row in songs.iloc[:10,:].iterrows():
  features = librosa_pipeline(row['track_id'])
  librosa_features.append(features)

end = time.time()

print(end-start)


librosa_features = []

start_1 = time.time()
for i,row in songs.iloc[:50,:].iterrows():
  features = delayed(librosa_pipeline)(row['track_id'])
  librosa_features.append(features)
end_1 = time.time()
print(end_1 - start_1)

start = time.time()
results = compute(*librosa_features)
end = time.time()
print(end-start)

librosa_features = []
start_1 = time.time()
for i,row in songs.iloc[50:,:].iterrows():
  features = delayed(librosa_pipeline)(row['track_id'])
  librosa_features.append(features)
end_1 = time.time()
print(end_1 - start_1)

from dask.diagnostics import ProgressBar

start = time.time()
with ProgressBar():
  rest_of_results = compute(*librosa_features, scheduler="processes")
end = time.time()

print(end-start)

# `results` contains the first 50 songs
# need to extend `results` with `rest_of_results`

len(rest_of_results)

results_df = pd.DataFrame(results)
results_df

rest_of_results_df = pd.DataFrame(rest_of_results)
rest_of_results_df

combined_df = results_df.append(rest_of_results_df,ignore_index=True)
combined_df

combined_df.to_csv('librosa_features.csv')
