==> notebooks/6-finalize-dataset.ipynb <==
## Finalize the Dataset
Before I jump into analysis, I need to create a final version of the dataset.

import pandas as pd
import numpy as np

# bring in the datasets
songs = pd.read_csv("../data/interim/all_decades_songs_V3.csv", index_col=0)
librosa_features = pd.read_csv("../data/processed/librosa_features.csv", index_col=0)
spotify_features = pd.read_csv("../data/processed/spotify_features.csv", index_col=0)

for df in [songs, librosa_features, spotify_features]:
    print(df.shape)

All the datasets have the same number of records, now I can decide which features to keep.

songs.columns

spotify_features['key'].value_counts()

# removing columns from `spotify_features`
cols = list(spotify_features.columns)

for x in ['type','uri','track_href','analysis_url']:
    cols.remove(x)

spotify_features_final  = spotify_features[cols]
spotify_features_final

librosa_features_df = librosa_features[['track_id','zero_crossing_rate']]
librosa_features_df

### Combine the datasets

songs.drop_duplicates(inplace=True)

songs

songs_df = songs.copy()

combined = songs_df.join(spotify_features.set_index('id'), on='track_id', how='left')
combined

combined.duplicated().value_counts()

combined.drop_duplicates(inplace=True)

combined

combined_final = combined.join(librosa_features.set_index('track_id'), on='track_id', how='left')
combined_final

combined_final.duplicated().value_counts()

combined_final.drop_duplicates(inplace=True)

combined_final.columns

combined_final_all_features = combined_final[['track_id', 'decade', 'track_name', 'artist_name', 
                           'danceability', 'energy', 'key', 'loudness', 
                           'mode', 'speechiness', 'acousticness', 'instrumentalness',
                           'liveness', 'valence', 'tempo','duration_ms','time_signature',
                           'spectral_centroid', 'spectral_bandwidth', 'rolloff', 
                           'zero_crossing_rate', 'mfcc1', 'mfcc2','mfcc3', 'mfcc4', 
                           'mfcc5', 'mfcc6', 'mfcc7', 'mfcc8', 'mfcc9', 'mfcc10',
                           'mfcc11', 'mfcc12', 'mfcc13', 'mfcc14', 'mfcc15', 'mfcc16', 'mfcc17',
                           'mfcc18', 'mfcc19', 'mfcc20', 'C', 'C#', 'D', 'D#', 'E', 'F', 'F#', 'G',
                           'G#', 'A', 'A#', 'B','genres']]

combined_final_all_features.to_csv('../data/processed/final_dataset_all_features.csv')

combined_final.to_csv('../data/final_dataset.csv')

### Finalizing the Dataset Pt. II

final_dataset = pd.read_csv("../data/processed/final_dataset.csv", index_col=0)
year_key = pd.read_csv("../data/interim/year_key.csv", index_col=0)
year_key

final_dataset.shape

final_dataset['year'] = year_key['date']
final_dataset

final_dataset.to_csv('../data/processed/final_dataset_year.csv')
